+++ packages/legend/src/Legend.js
@@ -1,6 +1,6 @@
 import React, { useRef, useEffect, useMemo } from 'react';
 import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core';
+import { makeStyles } from '@vitessce/styles';
 import { capitalize, getDefaultColor, cleanFeatureId } from '@vitessce/utils';
 import { select } from 'd3-selection';
 import { scaleLinear } from 'd3-scale';
@@ -10,11 +10,10 @@ import { isEqual } from 'lodash-es';
 import { getXlinkHref } from './legend-utils.js';
 
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles()(() => ({
   legend: {
     top: '2px',
     right: '2px',
-    zIndex: '100',
     fontSize: '10px !important',
     flexDirection: 'column',
     backgroundColor: 'rgba(215, 215, 215, 0.7)',
@@ -54,17 +53,23 @@ const rectMarginX = 2;
 
 function combineExtents(extents, featureAggregationStrategy) {
   if (Array.isArray(extents)) {
-    if (featureAggregationStrategy === 'first') {
-      return extents[0];
-    } if (featureAggregationStrategy === 'last') {
-      return extents.at(-1);
-    } if (typeof featureAggregationStrategy === 'number') {
-      const i = featureAggregationStrategy;
-      return extents[i];
-    } if (featureAggregationStrategy === 'sum') {
-      return extents.reduce((a, h) => [a[0] + h[0], a[1] + h[1]]);
-    } if (featureAggregationStrategy === 'mean') {
-      return extents.reduce((a, h) => [a[0] + h[0], a[1] + h[1]]).map(v => v / extents.length);
+    if (Array.isArray(extents?.[0])) {
+      // Extents is an array of [min, max] tuples.
+      if (featureAggregationStrategy === 'first') {
+        return extents[0];
+      } if (featureAggregationStrategy === 'last') {
+        return extents.at(-1);
+      } if (typeof featureAggregationStrategy === 'number') {
+        const i = featureAggregationStrategy;
+        return extents[i];
+      } if (featureAggregationStrategy === 'sum') {
+        return extents.reduce((a, h) => [a[0] + h[0], a[1] + h[1]]);
+      } if (featureAggregationStrategy === 'mean') {
+        return extents.reduce((a, h) => [a[0] + h[0], a[1] + h[1]]).map(v => v / extents.length);
+      }
+    } else {
+      // Extents is a single [min, max] tuple.
+      return extents;
     }
   }
   return null;
@@ -130,7 +135,7 @@ export default function Legend(props) {
   } = props;
 
   const svgRef = useRef();
-  const classes = useStyles();
+  const { classes } = useStyles();
 
   const isDarkTheme = theme === 'dark';
   const isStaticColor = obsColorEncoding === 'spatialChannelColor' || obsColorEncoding === 'spatialLayerColor';