+++ packages/utils/spatial-utils/src/spatial.js
@@ -95,10 +95,11 @@ export function physicalSizeToMatrix(xSize, ySize, zSize, xUnit, yUnit, zUnit) {
 
   // TODO: is this still needed
   // sizes are special objects with own equals method - see `unit` in declaration
-  if (!sizes[0].equals(sizes[1])) {
-    // Handle scaling in the Y direction for non-square pixels
-    scale[1] = divide(sizes[1], sizes[0]);
-  }
+  // This messess the dimensions when the x & y are little different (e.g., for the geomx data)
+  // if (!sizes[0].equals(sizes[1])) {
+  //   // Handle scaling in the Y direction for non-square pixels
+  //   scale[1] = divide(sizes[1], sizes[0]);
+  // }
   // END TODO: is this still needed
 
   // Add in z dimension needed for Matrix4 scale API.
@@ -459,6 +460,21 @@ export function hexToRgb(hex) {
     parseInt(result[2].toLowerCase(), 16),
     parseInt(result[3].toLowerCase(), 16),
   ];
+}
+
+
+// TODO: is this needed?
+// In the spatialdata metadata the axis name/type/unit info are also listed in
+// coordinateTransformations[].input|output.axes[] entries.
+export function normalizeAxes(axes) {
+  // Normalize axes to OME-NGFF v0.4 format.
+  return axes.map((axisInfo) => {
+    if (typeof axisInfo === 'string') {
+      // If the axis is a string, assume it is a name and set type to 'space'.
+      return { name: axisInfo, type: 'space' };
+    }
+    return axisInfo;
+  });
 }
 
 /**