+++ packages/view-types/link-controller/src/LinkController.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-console */
 import React, { useMemo, useEffect, useRef, useCallback, useState } from 'react';
+import { shallow } from 'zustand/shallow';
 import {
   useViewConfigStoreApi,
   createLoaders,
@@ -7,7 +8,7 @@ import {
 import {
   Checkbox,
   Grid,
-} from '@material-ui/core';
+} from '@vitessce/styles';
 
 import { log } from '@vitessce/globals';
 
@@ -78,15 +79,16 @@ export default function LinkController(props) {
   }, [linkID, linkEndpoint]);
 
   useEffect(() => viewConfigStoreApi.subscribe(
+    state => ({
+      viewConfig: state.viewConfig,
+      mostRecentConfigSource: state.mostRecentConfigSource,
+    }),
     ({ viewConfig, mostRecentConfigSource }) => {
       if (onConfigChange && viewConfig && mostRecentConfigSource === 'internal') {
         onConfigChange(viewConfig);
       }
     },
-    state => ({
-      viewConfig: state.viewConfig,
-      mostRecentConfigSource: state.mostRecentConfigSource,
-    }),
+    { equalityFn: shallow },
   ), [viewConfigStoreApi, onConfigChange]);
 
 
@@ -138,13 +140,13 @@ export default function LinkController(props) {
           To join the same session navigate to <a href="https://vitessce.link">https://vitessce.link</a> and enter the <b>Link ID</b> displayed here in the view.
           The session is synced as long as the <b>Link Active</b> Checkbox is activated.
         </p>
-        <Grid container direction="row" style={{ gridGap: '10px' }}>
-          <Grid item xs={5} style={{ whiteSpace: 'nowrap' }}>
+        <Grid container direction="row" sx={{ gridGap: '10px' }}>
+          <Grid size={5} sx={{ whiteSpace: 'nowrap' }}>
             <p style={{ fontSize: '25px' }}>Link ID:&nbsp;&nbsp;<b>{linkID}</b></p>
           </Grid>
-          <Grid item xs={5} style={{ fontSize: '25px', whiteSpace: 'nowrap' }}>
+          <Grid size={5} sx={{ fontSize: '25px', whiteSpace: 'nowrap' }}>
             Link Active: <Checkbox
-              style={{ marginTop: '-2px', marginLeft: '-10px' }}
+              sx={{ marginTop: '-2px', marginLeft: '-10px' }}
               color="primary"
               checked={sync}
               onChange={e => setSync(e.target.checked)}